refactor: move IsCUID from cloud/organization to pkg/util
Review feedback: cloud/env was reaching cross-package into
cloud/organization for a generic string check.
